requirements.txt
autopep8==1.5.3
boto3==1.14.24
botocore==1.17.24
docutils==0.15.2
jmespath==0.10.0
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
PyYAML==5.3.1
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
